pyproject.toml
[project]
name = "book-writer"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "beautifulsoup4>=4.14.3",
    "bitsandbytes>=0.49.0",
    "chromadb>=1.3.7",
    "ebooklib>=0.20",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "tiktoken>=0.12.0",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
